Added api services related to item

+++ ShareThingsToEarnBE/controllers/item.js
@@ -1,7 +1,13 @@
 const models =  require('../models');
 const Op = require('sequelize').Op;
+const { BlobServiceClient } = require('@azure/storage-blob');
 
-async function createItem(data, userId) {
+// Azure Blob Storage setup
+const AZURE_STORAGE_CONNECTION_STRING = process.env.AZURE_STORAGE_CONNECTION_STRING;
+const blobServiceClient = BlobServiceClient.fromConnectionString(AZURE_STORAGE_CONNECTION_STRING);
+const containerClient = blobServiceClient.getContainerClient('sharethingsearbimagedb');
+
+async function createItem(data, fileData, userId) {
     try {
         const userExists = await models.User.findByPk(userId);
         if(!userExists) throw new Error('User not exists');
@@ -13,7 +19,26 @@ async function createItem(data, userId) {
             ...data,
             userId
         });
-        return Item;
+
+        let itemImage = {
+            id: null
+        }
+        if (fileData) {
+            const blockBlobClient = containerClient.getBlockBlobClient(fileData.originalname);
+            const uploadBlobResponse = await blockBlobClient.uploadFile(fileData.path);
+            // Get the URL of the uploaded file
+            const documentUrl = blockBlobClient.url;
+            itemImage = await models.ItemImages.create({
+                itemId: Item.id,
+                imageUrl: documentUrl,
+                fileName: fileData.originalname
+            });
+        }
+
+        return {
+            itemImageId: itemImage.id,
+            Item: Item
+        };
     } catch(error) {
         console.log(error);
         throw error;
@@ -30,6 +55,10 @@ async function listItems(userId) {
                 {
                     model: models.Category,
                     attributes: ['id', 'name']
+                },
+                {
+                    model: models.ItemImages,
+                    attributes: ['imageUrl', 'fileName']
                 }
             ]
         })
@@ -42,13 +71,35 @@ async function listItems(userId) {
     }
 }
 
-async function updateItem(itemId, data, customerId) {
+async function updateItem(itemId, data, fileData, customerId) {
     try {
         const userExists = await models.User.findByPk(customerId);
         if(!userExists) throw new Error('User not exists');
         const ItemExists = await models.Item.findOne({where: {id: itemId}});
         if(!ItemExists) throw new Error('Item not exists');
+        const itemRequestExists = await models.ItemRequest.findOne({where: { itemId: itemId, status: { [Op.in]: ['pending', 'approved'] }}});
+        if(itemRequestExists) throw new Error('Item rented cannot be edited');
         await models.Item.update(data, { where: { id: itemId } })
+        if (fileData) {
+            const itemImageExists = await models.ItemImages.findOne({where: {itemId: itemId}});
+
+            const blockBlobClient = containerClient.getBlockBlobClient(fileData.originalname);
+            const uploadBlobResponse = await blockBlobClient.uploadFile(fileData.path);
+            // Get the URL of the uploaded file
+            const documentUrl = blockBlobClient.url;
+            if(itemImageExists) {
+                await models.ItemImages.update({
+                    imageUrl: documentUrl,
+                    fileName: fileData.originalname
+                }, { where: { id: itemImageExists.id } });
+            } else {
+                await models.ItemImages.create({
+                    itemId: itemId,
+                    imageUrl: documentUrl,
+                    fileName: fileData.originalname
+                })
+            }
+        }
         return {
             itemId,
             name: data.name
@@ -63,6 +114,10 @@ async function destroyItem(itemId) {
     try {
         const itemExists = await models.Item.findOne({where: {id: itemId}});
         if(!itemExists) throw new Error('Item not exists');
+        const itemRequestExists = await models.ItemRequest.findOne({where: { itemId: itemId, status: { [Op.in]: ['pending', 'approved'] }}});
+        if(itemRequestExists) throw new Error('Item rented cannot be destroyed');
+        await models.ItemRequest.destroy({ where: { itemId }});
+        await models.ItemImages.destroy({ where: { itemId }})
         await models.Item.destroy({ where: {id: itemId }});
         return {
             id: itemId
@@ -87,19 +142,32 @@ async function listItemsToBook(userId) {
     try {
         const userExists = await models.User.findByPk(userId);
         if(!userExists) throw new Error('User not exists');
+        if (userExists.phoneNumber === null) throw new Error('Please update your profile to proceed further');
+        const today = new Date();
+        today.setUTCHours(0, 0, 0, 0);
         const items = await models.Item.findAndCountAll({
             where: { userId: {
                 [Op.ne]: userId
-            }},
+            }, availabilityEndDate: {
+                [Op.gt]: today
+            }, isRequested: false
+        },
             include: [
                 {
                     model: models.Category,
                     attributes: ['id', 'name']
+                },
+                {
+                    model: models.User,
+                    attributes: ['phoneNumber', 'userName', 'email', 'address', 'country', 'pinCode']
+                },
+                {
+                    model: models.ItemImages,
+                    attributes: ['imageUrl', 'fileName']
                 }
             ]
         })
-        console.log(userExists)
-        if (userExists.phoneNumber === null) throw new Error('Please update your profile to proceed further');
+        
         return items;
     } catch(error) {
         console.log(error);
